Crate libslide

Modules

bignum

evaluator_rules

grammar

math

parser

partial_evaluator

scanner

utils

Structs

Bignum

EvaluatorContext

Functions

evaluate
Evaluates an expression to as simplified a form as possible. The evaluation may be partial, as some values (like variables) may be unknown.

parse_expression

parse_expression_pattern

scan