Module libslide::diagnostics

libslide's diagnostic module.

libslide does not emit user-facing diagnostic information itself, so the diagnostics returned by libslide should be

as complete as possible, so that a consumer can process as little or as much information as they want

easily transformable into some output form by downstream customers (namely the slide app)

Structs

AssociatedDiagnostic
A secondary diagnostic associated with a primary Diagnostic.

Diagnostic
A diagnostic for slide source code.

Enums

DiagnosticKind
The kind of a slide diagnostic.

Traits

DiagnosticRecord
Describes a code and detailed explanation for a diagnostic.

DiagnosticRegistry
Describes an individual registry of slide diagnostics.