Module libslide::emit

Emit strategies for the libslide grammar IR.

Structs

EmitConfig
Configuration options for emitting a slide grammar.

Enums

EmitFormat
The format in which a slide grammar should be emitted.

Traits

Emit
Implements the emission of a type in an EmitFormat.

Functions

emit_latex

emit_pretty

join_emits

vert_lines