Struct libslide::grammar::mem::RcExpr

pub struct RcExpr {
    expr: Rc<Expr>,
    pub(crate) span: Span,
}

An interned version of an expression.

NB: interned expressions are equivalent if they point to the same underlying expression, even though two interned expressions may have different spans.

Fields
expr: Rc<Expr>

The underlying expression.

span: Span

The original span of this expression from an input source code. Even though the expression is interned, this span is distinct and serves as a backwards-mapping to where the expression originally came from.

Implementations

impl RcExpr

pub(crate) fn new<Sp>(expr: Expr, span: Sp) -> Self where
Sp: Into<Span>,

Creates a new reference-counted expression at a span.