Enum libslide::grammar::statement::Stmt

pub enum Stmt {
    Expr(RcExpr),
    Assignment(Assignment),
}

A statement in a slide program.

Variants

Expr(RcExpr)

An expression statement is a statement that consists solely of an expression. For example, the slide program

1 + 1

contains one statement, that statement also being an expression.

Assignment(Assignment)

An assignment binds some value to a variable. For example the statement

x = 1 + 1

binds the expression "1 + 1" to "x".