Module libslide::linter

Provides linter-like diagnostics for a slide program.

Modules

expr_pat
Lints for an expression pattern in a slide program.

stmt
Lints for a statement in a slide program.

Structs

LintConfig
Describes the configuration to use when linting a slide grammar.

Traits

LintRule
Describes a slide program linter. A Linter is implemented on a slide Grammar.

Functions

lint_expr_pat
Lints a slide expression pattern.

lint_stmt
Lints a slide statement list.