Module libslide::math

A collection of algorithms used by evaluation rules.

This module is decoupled from evaluation rules and the [partial evaluator] because:

math's algorithms often use representations differing from the libslide grammar, for which transforming shims are required.

math can be developed independently from libslide, with the goal of eventual use in evaluator rules.

Re-exports

pub use gcd::*;

pub use poly::*;

pub use gcd_poly_zz::*;

Modules

fraction

gcd

gcd_poly_zz

poly