Module libslide::math::fraction

Structs

Dec2FracError

Functions

dec2frac
Converts a decimal number to its irreducible fractional representation by walking the Stern-Brocot tree. This is equivalent to a binary search of the Farey sequence.

Type Definitions

Fraction
Represents a fraction, consisting of a numerator and denominator.