Module libslide::math::gcd_poly_zz

Functions

gcd_interpolate
Interpolates step-wise gcd of h and x into a polynomial.

gcd_poly_zz_heu
Calculates the GCD of two polynomials, f and g, in ZZ (integer, integer) space using the GCDHEU heuristic algorithm (sources: 1, 2).

poly_coeffs_gcd
Returns the GCD of a polynomial's term coefficients.

poly_extract_common
Extracts a constant coefficient from two polynomials. Returns a tuple of (constant, f_quotient, g_quotient).

trivial_gcd
Handles trivial polynomial GCD cases, namely if one polynomial is empty.