Module libslide::parser::errors

Diagnostic errors produced by the parser.

Structs

ExpectedExpr
This error fires on token sequences that are expected to parse as an expression, but do not.

ExtraTokens
This error fires on tokens that are not connected to the rest of a primary statement in a slide program.

IllegalPattern
Patterns are illegal in a "regular" slide program; i.e. a program including a standard expression.

IllegalVariable
Variables are illegal in a slide expression pattern.

MismatchedClosingDelimiter
All opening delimiters with closing pairs must have that closing delimiter as a correctly-ordered complement in a slide program. In particular,

ParseErrors

UnmatchedClosingDelimiter
All closing delimiters with opening pairs must have that opening delimiter as a complement in a slide program. In particular,