Function libslide::partial_evaluator::evaluate_expr

pub fn evaluate_expr(
    expr: RcExpr, 
    rules: &[Rule], 
    ctxt: &ProgramContext
) -> RcExpr

Evaluates an expression to as simplified a form as possible. The evaluation may be partial, as some values (like variables) may be unknown. The returned expression is normalized.