Struct libslide::partial_evaluator::variable_expand::EagerVariableExpander

struct EagerVariableExpander<'a> {
    expr: RcExpr,
    expand_def: Option<&'a Assignment>,
}

Eagerly expands variables in an expression.

An EagerVariableExpander expands all matching variables on every call to expand.

For example, "a + b".expand("a = b + 1").expand("b = c") would expand to "c + 1 + c". Compare this to a LazyVariableExpander.

Note, however, that an EagerVariableExpander will not try to expand variables within an expanded variable definition on the same call to expand.

For example, "a".expand("a = a + a") would expand to "a + a" rather than "a + a + a + a + ...".

Fields
expr: RcExprexpand_def: Option<&'a Assignment>