Trait libslide::partial_evaluator::variable_expand::VariableExpander

trait VariableExpander<'a> {
    fn new(expr: RcExpr) -> Self;

    fn expand(self, asgn: &'a Assignment) -> Self;

    fn finish(self) -> RcExpr;
}

Required methods

fn new(expr: RcExpr) -> Self

Creates a new VariableExpander from an expression to expand.

fn expand(self, asgn: &'a Assignment) -> Self

Expands the variables matching the lhs of asgn with the assignment definition. Consumes and returns self, providing a chaining API.

fn finish(self) -> RcExpr

Consumes self and returns the owned expression with any variables expanded.

Loading content...