Struct libslide::scanner::Scanner

struct Scanner {
    pos: usize,
    input: Vec<char>,
    leading_trivia_start: usize,
    pub output: Vec<Token>,
    pub diagnostics: Vec<Diagnostic>,
}

Fields
pos: usizeinput: Vec<char>leading_trivia_start: usizeoutput: Vec<Token>diagnostics: Vec<Diagnostic>
Implementations

impl Scanner

pub fn new(input: &str) -> Scanner

Instantiate a new scanner.

fn peek(&self) -> Option<&char>

fn next(&mut self) -> Option<&char>

fn push_diag(&mut self, diagnostic: Diagnostic)

fn push_tok<S: Into<Span>>(&mut self, ty: TokenType, span: S)

fn collect_while(&mut self, pred: fn(_: &char) -> bool) -> String

pub fn scan(&mut self)

fn scan_trivia(&mut self)

Scans leading trivia, including whitespace.

fn scan_symbol(&mut self)

Matches a symbol with a token and creates it.

fn scan_num(&mut self)

Scans through the content of a number to create a token of that value.

fn scan_var_str(&mut self) -> String

fn scan_var(&mut self)

fn scan_var_pattern(&mut self)

fn scan_const_pattern(&mut self)

fn scan_any_pattern(&mut self)