Trait libslide::Emit

pub trait Emit where
    Self: Display + Debug, {
    fn emit_pretty(&self, config: EmitConfig) -> String;

    fn emit_s_expression(&self, config: EmitConfig) -> String;

    fn emit_latex(&self, config: EmitConfig) -> String;

    fn emit(&self, form: EmitFormat, config: EmitConfig) -> String { ... }

    fn emit_debug(&self, _config: EmitConfig) -> String { ... }

    fn emit_wrapped_latex(&self, config: EmitConfig) -> String { ... }
}

Implements the emission of a type in an [EmitFormat][EmitFormat].

Required methods

fn emit_pretty(&self, config: EmitConfig) -> String

Emit self with the [pretty emit format][EmitFormat::Pretty]

fn emit_s_expression(&self, config: EmitConfig) -> String

Emit self with the [s_expression emit format][EmitFormat::SExpression]

fn emit_latex(&self, config: EmitConfig) -> String

Emit self with the [LaTeX emit format][EmitFormat::Latex]

Loading content...
Provided methods

fn emit(&self, form: EmitFormat, config: EmitConfig) -> String

Emit self with the given [EmitFormat][EmitFormat].

NB: This is a multiplexer of the corresponding emit_ methods present on [Emit][Emit], except for [EmitFormat::Latex][EmitFormat::Latex], which is emitted via [emit_wrapped_latex][Emit::emit_wrapped_latex].

fn emit_debug(&self, _config: EmitConfig) -> String

Emit self with the [debug emit format][EmitFormat::Debug]

fn emit_wrapped_latex(&self, config: EmitConfig) -> String

Same as [emit_latex][Emit::emit_latex], but wraps the latex code in inline math mode.

Loading content...